+++ frontend/src/scenes/max/components/ContextDisplay.tsx
