+++ frontend/src/scenes/max/components/QuestionInput.scss
